upgrade: android15-6.6 in 2024/03 [2/2]
PD#SWPL-177341

Problem:
record merge upstream information

Solution:
AML_KERNEL_VERSION      6
UPSTREAM_VERSION        "a8759911ea10"
AML_PATCH_VERSION       "a8759911ea10"
MERGE_DATE              "2024-07-16"

Verify:
a4/a5

+++ drivers/media/common/dma-buf-heaps/amlogic_heap_cma.c
@@ -292,8 +292,8 @@ static const struct dma_buf_ops meson_cma_heap_buf_ops = {
 
 static struct dma_buf *meson_cma_heap_allocate(struct dma_heap *heap,
 					 unsigned long len,
-					 unsigned long fd_flags,
-					 unsigned long heap_flags)
+					 u32 fd_flags,
+					 u64 heap_flags)
 {
 	struct meson_cma_heap *meson_cma_heap = dma_heap_get_drvdata(heap);
 	struct meson_cma_heap_buffer *buffer;
