+++ drivers/media/common/dma-buf-heaps/amlogic_heap_secure.c
@@ -262,8 +262,8 @@ free_buffer:
 static struct dma_buf *meson_secure_uncached_heap_allocate
 						(struct dma_heap *heap,
 						unsigned long len,
-						unsigned long fd_flags,
-						unsigned long heap_flags)
+						u32 fd_flags,
+						u64 heap_flags)
 {
 	return meson_secure_heap_allocate(heap, len, fd_flags, heap_flags, true);
 }
@@ -271,8 +271,8 @@ static struct dma_buf *meson_secure_uncached_heap_allocate
 static struct dma_buf *meson_secure_uncached_heap_not_initialized
 						(struct dma_heap *heap,
 						unsigned long len,
-						unsigned long fd_flags,
-						unsigned long heap_flags)
+						u32 fd_flags,
+						u64 heap_flags)
 {
 	return ERR_PTR(-EBUSY);
 }
